+++ batch.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# pylint: disable=wrong-import-position,import-error,too-many-arguments
+# pylint: disable=import-error,no-value-for-parameter,too-many-arguments,wrong-import-position
 
 """
 Hail Batch workflow for the rare-variant association analysis, including:
@@ -45,6 +45,8 @@ from csv import DictReader
 
 import hail as hl
 import hailtop.batch as hb
+
+from _utils import qv_estimate as qvalue  # syntax?
 
 from cellregmap import (  # figure out how to import this from github
     run_gene_set_association,
@@ -564,129 +566,6 @@ def remove_sc_outliers(df, outliers=['966_967', '88_88']):
     return df
 
 
-def qvalue(pv, m=None, verbose=False, lowmem=False, pi0=None):
-    """
-    Estimates q-values from p-values
-    Args
-    =====
-    m: number of tests. If not specified m = pv.size
-    verbose: print verbose messages? (default False)
-    lowmem: use memory-efficient in-place algorithm
-    pi0: if None, it's estimated as suggested in Storey and Tibshirani, 2003.
-         For most GWAS this is not necessary, since pi0 is extremely likely to be
-         1
-
-    Taken from https://github.com/nfusi/qvalue
-    but updated for python 3 (e.g., xrange -> range)
-
-    Copyright (c) 2012, Nicolo Fusi, University of Sheffield
-    All rights reserved.
-
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions are met:
-        * Redistributions of source code must retain the above copyright
-        notice, this list of conditions and the following disclaimer.
-        * Redistributions in binary form must reproduce the above copyright
-        notice, this list of conditions and the following disclaimer in the
-        documentation and/or other materials provided with the distribution.
-        * Neither the name of the organization nor the
-        names of its contributors may be used to endorse or promote products
-        derived from this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-    ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-    WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER BE LIABLE FOR ANY
-    DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-    (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-    LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-    ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    """
-
-    assert pv.min() >= 0 and pv.max() <= 1, 'p-values should be between 0 and 1'
-
-    original_shape = pv.shape
-    pv = pv.ravel()  # flattens the array in place, more efficient than flatten()
-
-    if m is None:
-        m = float(len(pv))
-    else:
-        # the user has supplied an m
-        m *= 1.0
-
-    # if the number of hypotheses is small, just set pi0 to 1
-    if len(pv) < 100 and pi0 is None:
-        pi0 = 1.0
-    elif pi0 is not None:
-        pi0 = pi0
-    else:
-        # evaluate pi0 for different lambdas
-        pi0 = []
-        lam = sp.arange(0, 0.90, 0.01)
-        counts = sp.array([(pv > i).sum() for i in sp.arange(0, 0.9, 0.01)])
-        for l in range(len(lam)):
-            pi0.append(counts[l] / (m * (1 - lam[l])))
-
-        pi0 = sp.array(pi0)
-
-        # fit natural cubic spline
-        tck = interpolate.splrep(lam, pi0, k=3)
-        pi0 = interpolate.splev(lam[-1], tck)
-        if verbose:
-            # print('qvalues pi0=%.3f, estimated proportion of null features ' % pi0)
-            logging.info(
-                'qvalues pi0=%.3f, estimated proportion of null features ' % pi0
-            )
-
-        if pi0 > 1:
-            if verbose:
-                # print(
-                #     'got pi0 > 1 (%.3f) while estimating qvalues, setting it to 1' % pi0
-                # )
-                logging.info(
-                    'got pi0 > 1 (%.3f) while estimating qvalues, setting it to 1' % pi0
-                )
-            pi0 = 1.0
-
-    assert pi0 >= 0 and pi0 <= 1, 'pi0 is not between 0 and 1: %f' % pi0
-
-    if lowmem:
-        # low memory version, only uses 1 pv and 1 qv matrices
-        qv = sp.zeros((len(pv),))
-        last_pv = pv.argmax()
-        qv[last_pv] = (pi0 * pv[last_pv] * m) / float(m)
-        pv[last_pv] = -sp.inf
-        prev_qv = last_pv
-        for i in range(int(len(pv)) - 2, -1, -1):
-            cur_max = pv.argmax()
-            qv_i = pi0 * m * pv[cur_max] / float(i + 1)
-            pv[cur_max] = -sp.inf
-            qv_i1 = prev_qv
-            qv[cur_max] = min(qv_i, qv_i1)
-            prev_qv = qv[cur_max]
-
-    else:
-        p_ordered = sp.argsort(pv)
-        pv = pv[p_ordered]
-        qv = pi0 * m / len(pv) * pv
-        qv[-1] = min(qv[-1], 1.0)
-
-        for i in range(len(pv) - 2, -1, -1):
-            qv[i] = min(pi0 * m * pv[i] / (i + 1.0), qv[i + 1])
-
-        # reorder qvalues
-        qv_temp = qv.copy()
-        qv = sp.zeros_like(qv)
-        qv[p_ordered] = qv_temp
-
-    # reshape qvalues
-    qv = qv.reshape(original_shape)
-
-    return qv
-
-
 # endregion MISCELLANEOUS
 
 config = get_config()
@@ -831,13 +710,13 @@ def crm_pipeline(
             expression_files_prefix, 'expression_files', f'{celltype}_expression.tsv'
         )
 
-        genes = extract_genes(expression_files_prefix, genes_of_interest)
+        genes_list = extract_genes(expression_files_prefix, genes_of_interest)
         # maybe this makes no sense (to loop over genes twice)
         # but I also didn't want to re-select variants for the same gene repeatedly
         # for every new cell type?
         gene_prepare_jobs = []
         gene_run_jobs = []
-        for gene in genes:
+        for gene in genes_list:
             plink_output_prefix = gene_dict[gene]['plink']
             # prepare input files
             prepare_input_job = batch.new_python_job(f'Prepare inputs for: {gene}')
